Return scene vector s_1:n and event vector e_1:n in a format that's convenient for HRR encoding and passing to SEM.
Apologies that I'm using the world "scene" instead of "state". I'm
trying to stick with this nomenclature as "state" is an ambiguous term
that can stand for many other things in the model.

Also fixed a couple of small bugs:
- cond_fill is not always unicode and for me it was erroring out with
AttributeError: 'str' object has no attribute 'isnumeric'
- when GEN_SYMBOLIC_STATES and ATTACH_ROLE_MARKER are both set, the
role-filler pairs were printed multiple times (e.g. in the Dessert_crumble scene). I added
the vector used_fillers to keep track of the fillers independent of the
roles.